slib: Add compiler_details.h
This file was generated by configure, using a runtime test (which
prevents cross compiles). This uses information available to the
compiler already.

Later cleanup: it's a bit of a hack to reuse symbols that are
already defined in endian.h, so maybe rework the tree to use endian.h
idiomatically.